+++ client/client_test.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 	"testing"
 
-	_ "github.com/docker/libnetwork/netutils"
+	_ "github.com/docker/libnetwork/testutils"
 )
 
 // nopCloser is used to provide a dummy CallFunc for Cmd()
