+++ etchosts/etchosts_test.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"testing"
 
-	_ "github.com/docker/libnetwork/netutils"
+	_ "github.com/docker/libnetwork/testutils"
 )
 
 func TestBuildDefault(t *testing.T) {
