+++ idm/idm_test.go
@@ -3,7 +3,7 @@ package idm
 import (
 	"testing"
 
-	_ "github.com/docker/libnetwork/netutils"
+	_ "github.com/docker/libnetwork/testutils"
 )
 
 func TestNew(t *testing.T) {
